Use separate table props for projects table
