+++ src/index.scss
@@ -4,6 +4,7 @@
 body {
   margin: 0;
   font-size: 1.25rem;
+  padding-bottom: 3rem;
 }
 
 a {
