+++ src/pages/Projects.js
@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from 'react';
+import { NavLink } from 'react-router-dom';
 import axios from 'axios';
+import { useTable, usePagination, useSortBy } from 'react-table';
 import {
   cleanBrigadeData,
   getProjectsFromBrigadeData,
@@ -9,7 +11,42 @@ import ProjectsTable from '../components/ProjectsTable/ProjectsTable';
 
 function Projects() {
   const [brigadeData, setBrigadeData] = useState();
-  const projects = getProjectsFromBrigadeData(brigadeData);
+  const [projects, setProjects] = useState();
+
+  const columns = React.useMemo(
+    () => [
+      {
+        Header: 'Project name',
+        accessor: (project) => (
+          <NavLink to={`/projects/${project.slug}`}>{project.name}</NavLink>
+        ),
+        disableSortBy: true,
+        // TODO: TEXT FILTER
+      },
+      {
+        Header: 'Description',
+        accessor: 'description',
+        disableSortBy: true,
+        // TODO: TEXT FILTER
+      },
+      {
+        Header: 'Brigade',
+        accessor: 'brigade.name',
+        sortType: 'basic',
+      },
+    ],
+    []
+  );
+
+  const tableAttributes = useTable(
+    {
+      columns,
+      data: projects || [],
+      initialState: { pageIndex: 0, pageSize: projects ? projects.length : 50 },
+    },
+    useSortBy,
+    usePagination
+  );
 
   useEffect(() => {
     const getData = async () => {
@@ -19,11 +56,15 @@ function Projects() {
     getData();
   }, []);
 
+  useEffect(() => {
+    setProjects(getProjectsFromBrigadeData(brigadeData, {}));
+  }, [brigadeData]);
+
   return (
     <>
-      <h2>Projects page!</h2>
+      <h2>All projects</h2>
       {/* This is just a stand-in-- we should probably make it so that we can pass column props to the table */}
-      <ProjectsTable projects={projects} />
+      <ProjectsTable projects={projects} tableAttributes={tableAttributes} />
     </>
   );
 }
